(fleet) remove support for legacy OCI tar layouts
This PR removes support for downloading OCI tar layouts as they are now
fully replaced by OCI images downloaded directly from repositories.

commit-id:36899ce4
