+++ pkg/updater/download_test.go
@@ -10,12 +10,9 @@ package updater
 
 import (
 	"context"
-	"crypto/sha256"
 	"embed"
 	"fmt"
-	"io"
 	"io/fs"
-	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"os"
@@ -98,16 +95,14 @@ func buildOCIRegistry(t *testing.T) *httptest.Server {
 }
 
 type testFixturesServer struct {
-	t    *testing.T
-	s    *httptest.Server
-	soci *httptest.Server
+	t *testing.T
+	s *httptest.Server
 }
 
 func newTestFixturesServer(t *testing.T) *testFixturesServer {
 	return &testFixturesServer{
-		t:    t,
-		s:    httptest.NewServer(http.FileServer(http.FS(fixturesFS))),
-		soci: buildOCIRegistry(t),
+		t: t,
+		s: buildOCIRegistry(t),
 	}
 }
 
@@ -115,39 +110,15 @@ func (s *testFixturesServer) Downloader() *downloader {
 	return newDownloader(s.s.Client(), "")
 }
 
-func (s *testFixturesServer) DownloaderOCI() *downloader {
-	return newDownloader(s.soci.Client(), "")
-}
-
-func (s *testFixturesServer) DownloaderOCIRegistryOverride() *downloader {
-	return newDownloader(s.soci.Client(), "my.super/registry")
+func (s *testFixturesServer) DownloaderRegistryOverride() *downloader {
+	return newDownloader(s.s.Client(), "my.super/registry")
 }
 
 func (s *testFixturesServer) Package(f fixture) Package {
-	file, err := fixturesFS.Open(f.layoutPath)
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-	hash := sha256.New()
-	n, err := io.Copy(hash, file)
-	if err != nil {
-		panic(err)
-	}
 	return Package{
 		Name:    f.pkg,
 		Version: f.version,
-		URL:     s.s.URL + "/" + f.layoutPath,
-		Size:    n,
-		SHA256:  fmt.Sprintf("%x", hash.Sum(nil)),
-	}
-}
-
-func (s *testFixturesServer) PackageOCI(f fixture) Package {
-	return Package{
-		Name:    f.pkg,
-		Version: f.version,
-		URL:     fmt.Sprintf("oci://%s/%s@%s", strings.TrimPrefix(s.soci.URL, "http://"), f.pkg, f.indexDigest),
+		URL:     fmt.Sprintf("oci://%s/%s@%s", strings.TrimPrefix(s.s.URL, "http://"), f.pkg, f.indexDigest),
 	}
 }
 
@@ -171,8 +142,7 @@ func (s *testFixturesServer) ConfigFS(f fixture) fs.FS {
 }
 
 func (s *testFixturesServer) Image(f fixture) oci.Image {
-	tmpDir := s.t.TempDir()
-	image, err := s.Downloader().Download(context.Background(), tmpDir, s.Package(f))
+	image, err := s.Downloader().Download(context.Background(), s.Package(f))
 	if err != nil {
 		panic(err)
 	}
@@ -190,7 +160,6 @@ func (s *testFixturesServer) Catalog() catalog {
 
 func (s *testFixturesServer) Close() {
 	s.s.Close()
-	s.soci.Close()
 }
 
 func TestDownload(t *testing.T) {
@@ -198,7 +167,7 @@ func TestDownload(t *testing.T) {
 	defer s.Close()
 	d := s.Downloader()
 
-	image, err := d.Download(context.Background(), t.TempDir(), s.Package(fixtureSimpleV1))
+	image, err := d.Download(context.Background(), s.Package(fixtureSimpleV1))
 	assert.NoError(t, err)
 	tmpDir := t.TempDir()
 	err = extractPackageLayers(image, t.TempDir(), tmpDir)
@@ -211,7 +180,7 @@ func TestPackage(t *testing.T) {
 	defer s.Close()
 	d := s.Downloader()
 
-	expectedPackage := s.PackageOCI(fixtureSimpleV1)
+	expectedPackage := s.Package(fixtureSimpleV1)
 	pkg, err := d.Package(context.Background(), expectedPackage.URL)
 	assert.NoError(t, err)
 	assert.Equal(t, expectedPackage.Name, pkg.Name)
@@ -225,8 +194,8 @@ func TestDownloadInvalidHash(t *testing.T) {
 	d := s.Downloader()
 
 	pkg := s.Package(fixtureSimpleV1)
-	pkg.SHA256 = "2857b8e9faf502169c9cfaf6d4ccf3a035eccddc0f5b87c613b673a807ff6d23"
-	_, err := d.Download(context.Background(), t.TempDir(), pkg)
+	pkg.URL = pkg.URL[:strings.Index(pkg.URL, "@sha256:")] + "@sha256:2857b8e9faf502169c9cfaf6d4ccf3a035eccddc0f5b87c613b673a807ff6d23"
+	_, err := d.Download(context.Background(), pkg)
 	assert.Error(t, err)
 }
 
@@ -237,7 +206,7 @@ func TestDownloadInvalidVersion(t *testing.T) {
 
 	pkg := s.Package(fixtureSimpleV1)
 	pkg.Version = "v1-1"
-	_, err := d.Download(context.Background(), t.TempDir(), pkg)
+	_, err := d.Download(context.Background(), pkg)
 	assert.Error(t, err)
 }
 
@@ -247,29 +216,16 @@ func TestDownloadPlatformNotAvailable(t *testing.T) {
 	d := s.Downloader()
 
 	pkg := s.Package(fixtureSimpleV1Linux2Amd128)
-	_, err := d.Download(context.Background(), t.TempDir(), pkg)
+	_, err := d.Download(context.Background(), pkg)
 	assert.Error(t, err)
-}
-
-func TestDownloadRegistry(t *testing.T) {
-	s := newTestFixturesServer(t)
-	defer s.Close()
-	d := s.DownloaderOCI()
-
-	image, err := d.Download(context.Background(), t.TempDir(), s.PackageOCI(fixtureSimpleV1))
-	assert.NoError(t, err)
-	tmpDir := t.TempDir()
-	err = extractPackageLayers(image, t.TempDir(), tmpDir)
-	assert.NoError(t, err)
-	assertEqualFS(t, s.PackageFS(fixtureSimpleV1), os.DirFS(tmpDir))
 }
 
 func TestDownloadRegistryWithOverride(t *testing.T) {
 	s := newTestFixturesServer(t)
 	defer s.Close()
-	d := s.DownloaderOCIRegistryOverride()
+	d := s.DownloaderRegistryOverride()
 
-	_, err := d.Download(context.Background(), t.TempDir(), s.PackageOCI(fixtureSimpleV1))
+	_, err := d.Download(context.Background(), s.Package(fixtureSimpleV1))
 	assert.Error(t, err) // Host not found
 }
 
@@ -280,15 +236,15 @@ func TestGetRegistryURL(t *testing.T) {
 	pkg := Package{
 		Name:    "simple",
 		Version: "v1",
-		URL:     s.soci.URL + "/simple@sha256:2aaf415ad1bd66fd9ba5214603c7fb27ef2eb595baf21222cde22846e02aab4d",
+		URL:     s.s.URL + "/simple@sha256:2aaf415ad1bd66fd9ba5214603c7fb27ef2eb595baf21222cde22846e02aab4d",
 		SHA256:  "2aaf415ad1bd66fd9ba5214603c7fb27ef2eb595baf21222cde22846e02aab4d",
 	}
 
-	d := s.DownloaderOCI()
+	d := s.Downloader()
 	url := d.getRegistryURL(pkg.URL)
-	assert.Equal(t, s.soci.URL+"/simple@sha256:2aaf415ad1bd66fd9ba5214603c7fb27ef2eb595baf21222cde22846e02aab4d", url)
+	assert.Equal(t, s.s.URL+"/simple@sha256:2aaf415ad1bd66fd9ba5214603c7fb27ef2eb595baf21222cde22846e02aab4d", url)
 
-	d = s.DownloaderOCIRegistryOverride()
+	d = s.DownloaderRegistryOverride()
 	url = d.getRegistryURL(pkg.URL)
 	assert.Equal(t, "my.super/registry/simple@sha256:2aaf415ad1bd66fd9ba5214603c7fb27ef2eb595baf21222cde22846e02aab4d", url)
 }
@@ -296,9 +252,9 @@ func TestGetRegistryURL(t *testing.T) {
 func TestDownloadOCIPlatformNotAvailable(t *testing.T) {
 	s := newTestFixturesServer(t)
 	defer s.Close()
-	d := s.DownloaderOCI()
+	d := s.Downloader()
 
-	pkg := s.PackageOCI(fixtureSimpleV1Linux2Amd128)
-	_, err := d.Download(context.Background(), t.TempDir(), pkg)
+	pkg := s.Package(fixtureSimpleV1Linux2Amd128)
+	_, err := d.Download(context.Background(), pkg)
 	assert.Error(t, err)
 }
