+++ pkg/updater/install.go
@@ -6,9 +6,13 @@
 package updater
 
 import (
+	"archive/tar"
 	"fmt"
+	"io"
+	"io/fs"
 	"os"
 	"path/filepath"
+	"strings"
 
 	oci "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/google/go-containerregistry/pkg/v1/types"
@@ -143,6 +147,82 @@ func extractPackageLayers(image oci.Image, configDir string, packageDir string) 
 		default:
 			log.Warnf("can't install unsupported layer media type: %s", mediaType)
 		}
+	}
+	return nil
+}
+
+// extractTarArchive extracts a tar archive to the given destination path
+//
+// Note on security: This function does not currently attempt to fully mitigate zip-slip attacks.
+// This is purposeful as the archive is extracted only after its SHA256 hash has been validated
+// against its reference in the package catalog. This catalog is itself sent over Remote Config
+// which guarantees its integrity.
+func extractTarArchive(reader io.Reader, destinationPath string, maxSize int64) error {
+	log.Debugf("Extracting archive to %s", destinationPath)
+	tr := tar.NewReader(io.LimitReader(reader, maxSize))
+	for {
+		header, err := tr.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			return fmt.Errorf("could not read tar header: %w", err)
+		}
+		if header.Name == "./" {
+			continue
+		}
+
+		target := filepath.Join(destinationPath, header.Name)
+
+		// Check for directory traversal. Note that this is more of a sanity check than a security measure.
+		if !strings.HasPrefix(target, filepath.Clean(destinationPath)+string(os.PathSeparator)) {
+			return fmt.Errorf("tar entry %s is trying to escape the destination directory", header.Name)
+		}
+
+		// Extract element depending on its type
+		switch header.Typeflag {
+		case tar.TypeDir:
+			err = os.MkdirAll(target, os.FileMode(header.Mode))
+			if err != nil {
+				return fmt.Errorf("could not create directory: %w", err)
+			}
+		case tar.TypeReg:
+			err = extractTarFile(target, tr, os.FileMode(header.Mode))
+			if err != nil {
+				return err // already wrapped
+			}
+		case tar.TypeSymlink:
+			err = os.Symlink(header.Linkname, target)
+			if err != nil {
+				return fmt.Errorf("could not create symlink: %w", err)
+			}
+		case tar.TypeLink:
+			// we currently don't support hard links in the updater
+		default:
+			log.Warnf("Unsupported tar entry type %d for %s", header.Typeflag, header.Name)
+		}
+	}
+
+	log.Debugf("Successfully extracted archive to %s", destinationPath)
+	return nil
+}
+
+// extractTarFile extracts a file from a tar archive.
+// It is separated from extractTarGz to ensure `defer f.Close()` is called right after the file is written.
+func extractTarFile(targetPath string, reader io.Reader, mode fs.FileMode) error {
+	err := os.MkdirAll(filepath.Dir(targetPath), 0755)
+	if err != nil {
+		return fmt.Errorf("could not create directory: %w", err)
+	}
+	f, err := os.OpenFile(targetPath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, os.FileMode(mode))
+	if err != nil {
+		return fmt.Errorf("could not create file: %w", err)
+	}
+	defer f.Close()
+
+	_, err = io.Copy(f, reader)
+	if err != nil {
+		return fmt.Errorf("could not write file: %w", err)
 	}
 	return nil
 }
